pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shl-conversational-recommender"
version = "0.1.0"
description = "Conversational SHL assessment recommender built for the supplied catalog and sample conversations."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.136",
  "numpy>=1.26",
  "scikit-learn>=1.7",
  "python-dotenv>=1.0",
  "uvicorn>=0.46",
]

[project.optional-dependencies]
llm = [
  "langchain-core>=0.3",
  "langchain-openai>=0.2",
  "openai>=1.0",
]
dev = [
  "httpx>=0.28",
  "pytest>=8.4",
]

[tool.pytest.ini_options]
pythonpath = ["."]
